requirements.txt
python-telegram-bot>=20.0
yt-dlp>=2024.0.0
loguru>=0.7.0
